backspacestringcompare_test.go
package LeetCode

import "testing"

func Test_backspaceCompare(t *testing.T) {
	type args struct {
		s string
		t string
	}
	tests := []struct {
		name string
		args args
		want bool
	}{
		{"LeetCode #1", args{"ab#c", "ad#c"}, true},
		{"LeetCode #2", args{"ab##", "c#d#"}, true},
		{"LeetCode #3", args{"a#c", "b"}, false},
		{"Hard one", args{"j##yc##bs#srqpfzantto###########i#mwb", "j##yc##bs#srqpf#zantto###########i#mwb"}, false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := backspaceCompare(tt.args.s, tt.args.t); got != tt.want {
				t.Errorf("backspaceCompare() = %v, want %v", got, tt.want)
			}
		})
	}
}
